apps/requirements.txt
APScheduler
arabic-reshaper
asgiref
asn1crypto
bs4
certifi
cffi
chardet
charset-normalizer
click
cryptography
cssselect2
Django==4.2.21
django-auditlog
django-apscheduler
django-cors-headers
django-environ
django-filter
django-haystack
django-import-export
django-jsonfield
django-mathfilters
django-microsoft-auth
django-model-utils
django-simple-history
django-widget-tweaks
djangorestframework
djangorestframework-simplejwt
drf-yasg
et-xmlfile
geopy
google-api-python-client
google-cloud-storage==3.0.0
html5lib
Jinja2
idna
lxml
numpy
oauthlib
openpyxl
oscrypto
pandas
pdfkit
pillow
pycparser
pyHanko
pyhanko-certvalidator
PyMuPDF== 1.24.5
pypdf
pypng
python-bidi
python-dateutil
pytz
PyYAML
pyzk==0.9
qrcode
reportlab==4.0.9
requests
responses
setuptools
six
sqlparse
svglib
swapper
tinycss2
typing_extensions
tzdata
tzlocal
uritools
urllib3
utils
webencodings
Whoosh
xhtml2pdf
XlsxWriter
gunicorn
psycopg2-binary
whitenoise
django-storages

# RAG System and N8N Integration
chromadb>=0.4.15
langchain>=0.0.350
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
aiohttp>=3.8.0
PyPDF2>=3.0.1
python-docx>=0.8.11
python-pptx>=0.6.21
tqdm>=4.66.0
python-dotenv>=1.0.0
asyncio>=3.4.3
